requirements.txt
Markdown==3.3.6
matplotlib==3.5.0rc1
numpy==1.18.5
pandas==1.4.3
Pillow==8.4.0
pytest==4.6.9
PyYAML==5.3.1
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.4.1
torch==1.12.0
torchvision==0.13.0
typing-extensions==4.0.0
pyTelegramBotAPI==4.6.1

